move create_ddG_plot method to separate module for plotting (to be used by NEQ)
